fix: moved json and modified carousel.Item
